Fix ending mission due to not enough resource
r5076
2019-10-07

## CHANGE
1. Avoid loading issue with lookupPerson in Airlock.
   - Compare operatorID to whether it's > 0 in timePassing()
     in Airlock.
2. Add back equals() and hashCode() in Person, Equipment,
   Settlement, Building, Good.
3. Refine log statements in ExitAirlock and EnterAirlock.
4. Modify checkPersonStatus() to check if a person has
   any strength left when disembarking.
5. Use OR in isHungry() in PhysicalCondition since
   hunger and energy are not necessarily corresponding to
   each other in value.
6. Add the use of work gloves and decontamination kit in
   some malfunctions.
7. Increase the # of "rf power transistor" and
   "microcontroller" (they cannot be manufactured) in
   part_packages.xml.
   - Reduce their probability of malfunction.

## FIX
1. Correct getPerformanceStatus() in PhysicalCondition.
2. Correct getTotalRemainingDistance() in TravelMission.

+++ mars-sim-core/src/main/java/org/mars_sim/msp/core/person/ai/task/LoadVehicleEVA.java
@@ -786,11 +786,11 @@ public class LoadVehicleEVA extends EVAOperation implements Serializable {
 		Iterator<Integer> iE = optionalEquipment.keySet().iterator();
 		while (iE.hasNext() && (amountLoading > 0D)) {
 			Integer id = iE.next();
-			int numNeededTotal = (Integer) optionalEquipment.get(id);
+			int numNeededTotal = optionalEquipment.get(id);
 //			int i0 = numNeededTotal;
 			
 			if (requiredEquipment.containsKey(id)) {
-				numNeededTotal += (Integer) requiredEquipment.get(id);
+				numNeededTotal += requiredEquipment.get(id);
 			}
 			int numAlreadyLoaded = vInv.findNumUnitsOfClass(id);
 			
@@ -906,7 +906,7 @@ public class LoadVehicleEVA extends EVAOperation implements Serializable {
 			Iterator<Class<? extends Equipment>> i = equipment.keySet().iterator();
 			while (i.hasNext()) {
 				Class<? extends Equipment> equipmentType = i.next();
-				int num = (Integer) equipment.get(equipmentType);
+				int num = equipment.get(equipmentType);
 //				Coordinates defaultLoc = new Coordinates(0D, 0D);
 				for (int x = 0; x < num; x++) {
 					inv.storeUnit(EquipmentFactory.createEquipment(equipmentType, settlement.getCoordinates(), true));
